chore: separate Register into separate component
