+++ frontend/setup-tests.ts
@@ -1,0 +1,1 @@
+import '@testing-library/jest-dom'
