--- frontend/src/tests/setup.ts
@@ -1,2 +1,0 @@
-import '@testing-library/jest-dom'
-
